requirements.txt
fastapi
pydantic
python-dotenv
gspread
httpx
oauth2client
google-auth
google-auth-oauthlib
requests
faiss-cpu
PyMuPDF
